+++ src/userInsights.ts
